requirements.txt
streamlit
opencv-python
numpy
torch
torchvision
ultralytics
streamlit
inference-sdk
pillow